Python_packages/requirements.txt
Django~=2.2.4
IPython~=7.12.0
pytz
Pillow